backend/requirements.txt
absl-py
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
asgiref
astunparse
attrs
beautifulsoup4
bitarray
blinker
boto3
botocore
cached-property
certifi
cffi
charset-normalizer
ckzg
click
cryptography
cytoolz
deepface
Django
django-cors-headers
django-otp
django-rest-knox
django-rest-passwordreset
django-shortcuts
djangorestframework
djangorestframework_simplejwt
dlib
dynaconf
eth-account
eth-bloom
eth-hash
eth-keyfile
eth-keys
eth-rlp
eth-tester
eth-typing
eth-utils
eth_abi
face-recognition
face_recognition_models
filelock
fire
Flask
flask-cors
flatbuffers
frozenlist
gast
gdown
google-pasta
grpcio
gunicorn
h5py
hexbytes
hvac
idna
itsdangerous
Jinja2
jmespath
joblib
keras
knox
libclang
loguru
lru-dict
lz4
Markdown
markdown-it-py
MarkupSafe
mdurl
ml_dtypes
mtcnn
multidict
namex
numpy
opencv-python
opt_einsum
optree
packaging
pandas
parsimonious
pillow
propcache
protobuf
psycopg2
py-ecc
py-evm
py-geth
pycparser
pycryptodome
pydantic
pydantic_core
Pygments
PyJWT
pyotp
PySocks
pytesseract
python-dateutil
pytz
pyunormalize
regex
requests
retina-face
rich
rlp
s3transfer
safe-pysha3
semantic-version
setuptools
shortuuid
six
sortedcontainers
soupsieve
sqlparse
tensorboard
tensorboard-data-server
tensorflow
termcolor
tf_keras
toolz
tqdm
trie
types-requests
typing-inspection
typing_extensions
tzdata
urllib3
validators
web3
websockets
Werkzeug
wheel
wrapt
yarl